chg: processHandshake based on Connection
